Criando hook para o resumo
